requirements.txt
# Web Framework
Flask==3.0.3

# AI/ML
openai==1.12.0
google-generativeai==0.3.2 # for gemini

# HTTP Client
requests==2.31.0

# Utilities
python-dotenv==1.0.1
